feat: add syntax highlighting
